vector_service/src/requirements.txt
pinecone-client
openai
sqlalchemy
psycopg2-binary
python-dotenv
